backend/requirements.txt
SQLAlchemy==2.0.31
Flask==3.0.3
PyJWT~=2.8.0
werkzeug~=3.0.3
PyMySQL==1.1.1